Update to rand 0.6 &  a few other type changes

+++ src/render_settings.rs
@@ -2,7 +2,7 @@
 
 #[derive(Debug, Copy, Clone)]
 pub struct RenderSettings {
-    pub spp: u32,
+    pub spp: usize,
     pub max_depth: u32,
 }
